Add straight implementation of definitions
The json_schema function got an additional positional argument which
allows to store and access references during composition of a schema.

The refs argument could be None (which is set by default) or an instance
of the References class. If json_schema is called with refs=None and the
schema argument contains TypeRef objects, then the definitions item will
be added into the resulting schema with a value which will contain
schemas of TypeRef objects. If json_schema is called with an instance of
References, then the passed object will be used to store found
references during composition of a schema, hovewer the resulting schema
will not contain the definitions item and handling of definitions have
to be done explicetely.

This change can break backward compatibility in case if the json_schema
function was used for schemas which contain TypeRef objects and which
are not considered as top level schemas.